bot/requirements.txt
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.1
idna==3.3
Pillow==9.2.0
psycopg2-binary==2.9.3
pyTelegramBotAPI==4.7.0
Telethon==1.25.0
typing-extensions==4.3.0
urllib3==1.26.12
